+++ backend/model/Database.py
@@ -1,5 +1,5 @@
 from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import declaravite_base
+from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
 SQL_ALCHEMY_URL = "sqlite:///database.db"
@@ -7,6 +7,6 @@ SQL_ALCHEMY_URL = "sqlite:///database.db"
 engine = create_engine(
     SQL_ALCHEMY_URL, connect_args={"check_same_thread": False}
 )
-SessionLocal = sessionmaker(autocommit= False, autoFlush = False, bind = engine)
+SessionLocal = sessionmaker(autocommit= False, bind = engine)
 
-Base = declaravite_base()
+Base = declarative_base()
